requirements.txt
pytelegrambotapi
qrcode[pil]
pillow
python-dotenv